Move email templates to its own module
